move tests to more appropriate script
